requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework (for API)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Utilities
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
